Move image resource behind /service/v1
