+++ src/main/java/de/kja/server/Server.java
@@ -8,9 +8,9 @@ import de.kja.server.auth.DatabaseAuthenticator;
 import de.kja.server.dbi.AdminDao;
 import de.kja.server.dbi.ContentDao;
 import de.kja.server.dbi.DistrictDao;
-import de.kja.server.resources.images.ImagesResource;
 import de.kja.server.resources.service.ContentResource;
 import de.kja.server.resources.service.DistrictResource;
+import de.kja.server.resources.service.ImagesResource;
 import de.kja.server.resources.webinterface.EditContentResource;
 import de.kja.server.resources.webinterface.IndexResource;
 import io.dropwizard.Application;
